+++ lib/llm_cost_tracker/daily_total.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "active_record"
+
+module LlmCostTracker
+  class DailyTotal < ActiveRecord::Base
+    self.table_name = "llm_cost_tracker_daily_totals"
+  end
+end