+++ lib/llm_cost_tracker/railtie.rb
@@ -3,6 +3,7 @@
 module LlmCostTracker
   class Railtie < Rails::Railtie
     generators do
+      require_relative "generators/llm_cost_tracker/add_daily_totals_generator"
       require_relative "generators/llm_cost_tracker/add_monthly_totals_generator"
       require_relative "generators/llm_cost_tracker/add_latency_ms_generator"
       require_relative "generators/llm_cost_tracker/add_provider_response_id_generator"